packages/pi-controller/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
paho-mqtt>=2.0.0
RPi.GPIO>=0.7.1
psutil>=5.9.0
aiosqlite>=0.20.0
python-multipart>=0.0.9
requests>=2.31.0
